Add About Us and Product Detail pages with responsive design
- Created 'about.blade.php' to introduce Ibyora Beads, including sections for vision, mission, products, and customer reviews.
- Developed 'show.blade.php' for displaying individual product details, including image, name, color, and price.
- Implemented responsive layout and styling for better user experience across devices.

+++ database/migrations/2025_11_01_234408_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('color');
-            $table->decimal('price', 8, 2);
+            $table->integer('price');
             $table->string('image')->nullable();
             $table->timestamps();   
         });
